Fixed finding the next piece during low framerate.
When there's a lot framerate, the train jumps around a lot. We were just
using the train's current position to find the next track pice to use. However,
that caused problems when the train was still a long way away from the next
track piece. So now, we use the current traverser, get the last point, and
use that to find the next track piece.